need 1 script per region; start with AI
put gen.cols and gen.units in own scripts

created a template script

add dput() for data sourcing; @mpinsky: I should eventually use the actual data as sourced & merged by complete_r_script.R, but that's currently a very complicated and long script, so I'm avoiding rearranging that to get this to you quickly. #1 needs to be implemented if meta data are to be updated and/or associated with the actual data set.