Add raw data, except gmex
Within data_raw/ there are sub-directories for each region (including
some not used by OceanAdapt). Within each subdirectory, there is a zip
file for each download of the data from that region. The files are
named by the date of the download. Gmex is not included because the zip
file is 64MB.