Update download_scot.R
Updated download_scot.R with an additional location for the Scotian Shelf data. Mike McMahon said that he will directly upload the data to this location as soon as they are ready; however, they are in a different format than the data we usually use, and we will need to figure out how to use them.